% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterizePulp.R
\name{characterizePulp}
\alias{characterizePulp}
\title{Model selection for pulp characterization using C-Vine copulas}
\usage{
characterizePulp(pulpPath, saveDirName)
}
\arguments{
\item{pulpPath}{Path to the fiber characterization file}

\item{saveDirName}{Path to the place where files are saved}
}
\description{
characterizePulp() takes as input a dataframe
}
\examples{
characterizePulp()
characterizePulp
}
