% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFromMargin.R
\name{drawFromMargin}
\alias{drawFromMargin}
\title{Margin sampler}
\usage{
drawFromMargin(marginGenerator, quantilesToEval)
}
\arguments{
\item{marginGenerator}{Contains fitted univariate margin parameters}

\item{quantilesToEval}{Contains quantiles to invert to obtain value}
}
\description{
This function allows you to express your love of cats.
}
\examples{
drawFromMargin()

}
\keyword{cats}
