% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePseudoU.R
\name{generatePseudoU}
\alias{generatePseudoU}
\title{Pseudo-observation generator}
\usage{
generatePseudoU(observedSample)
}
\arguments{
\item{observedSample}{Matrix or dataframe of observations}
}
\description{
generatePseudoU() takes as input a matrix or dataframe of observations
and returns a matrix of ranked pseudo-observations
where d is the number of columns in the input.
}
\details{
The observations are normalized such that the value 1 is never
returned, as this could generate some numerical difficulties
later on.
}
\examples{
generatePseudoU()

}
