% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePulpSample.R
\name{generatePulpSample}
\alias{generatePulpSample}
\title{Generate a new sample given a complete data set}
\usage{
generatePulpSample(pulpPath, saveDirName, outputName, numSamples)
}
\arguments{
\item{pulpPath}{Path to the fiber characterization file}

\item{saveDirName}{Path to the place where the pre-computed copula is stored}

\item{outputName}{The name of the place where file is to be saved}

\item{numSamples}{The number of samples to generate}
}
\description{
generatePulpSample() takes as input a dataframe
}
\examples{
generatePulpSample()

}
