% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndFormatPulpFile.R
\name{readAndFormatPulpFile}
\alias{readAndFormatPulpFile}
\title{File input and restructuring}
\usage{
readAndFormatPulpFile(pulpPath)
}
\arguments{
\item{pulpPath}{Path to the fiber characterization file}
}
\description{
readAndFormatPulpFile() takes as input a path to a text file containing the raw fiber network,
and massages the shape so that it fits what the rest of the package expects
}
\examples{
readAndFormatPulpFile()

}
